changed types

+++ data/data.h
@@ -17,7 +17,7 @@ namespace Data {
   public:
     Imu imu;
     uint16_t distance_to_target;
-    Cartesian2<int32_t> position;
+    Cartesian2<float> position;
     uint8_t velocity;
   };
 }
